requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pillow==10.0.1
pypdf2==3.0.1
pdf2image==1.16.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
img2pdf==0.4.4
reportlab==4.0.4
markdown2==2.4.10
pydub==0.25.1
aiofiles==23.2.1
aiohttp==3.9.1
python-magic-bin==0.4.14
psutil==5.9.6
pywin32==306; sys_platform == 'win32'
pdf2docx==0.5.8
pdfplumber==0.10.3
ebooklib==0.18
python-pptx==0.6.23
ffmpeg-python==0.2.0